Use 'short' for short values

+++ com.io7m.jnoisetype.api/src/main/java/com/io7m/jnoisetype/api/NTInstrumentZoneModulatorType.java
@@ -42,7 +42,7 @@ public interface NTInstrumentZoneModulatorType extends NTInstrumentZoneMemberTyp
    */
 
   @NTNativeType(name = "short")
-  int modulationAmount();
+  short modulationAmount();
 
   /**
    * @return The degree to which the source modulates the destination is to be controlled by the
